ethernet: ti: am65-cpsw-qos: add TAPRIO (EST) configuration support
The iproute2 interface allows to configure EST cycle setting while
another one is still operational, and next new cycle - admin one,
starts right after old one is completed. So, it's convenient way to
split fetch RAM on two and work in two buffer operation mode.
Each buffer has 64 locations that should be enough and configurations
are supposed to be toggled w/o harm at next timer cycle. But it's not
only case, that can be implemented. If number of commands is not
enough to cover requested intervals for schedules, the one buffer mode
can to be used, but toggle can happen only if EST is stopped before.
Therefore the one buf mode is cut off for now and used only two buffer
mode for simplicity.

Also keep taprio configuration if link connection is lost for time
less then 1 second, by configuring EST with same parameters. In all
other cases stop EST and inform user about it.

Don't allow EST to be enabled if pf_p0_rx_ptype_rrobin privat flag is
set. The flag is common for all ports, so can't be just overridden by
taprio configuration w/o user involvement, thus, inform about
collision and don't allow the flag to be set if EST is active.

There is no way to set cycle time extension parameter, so ignore for
now. To implement it in software is also not trivial task as it
requires to find last cycle of oper configuration and then just stop
it before new cycle is scheduled, this time can be easily skipped, so
better block time extension feature for now.

Scheds (commands) configuration depends on interface speed so
link speed is used for that. Each schedule can be constructed with
several cmds depending on interval. For instance if each shced has
timer interval < ~130us on 1000 Mb link then each sched consumes one
command and have 1:1 mapping.

The patch allows to update the timer and scheds memory only if it's
really needed, and skip cases required the user to stop timer by
configuring only shceds memory.

Signed-off-by: Ivan Khoronzhuk <ivan.khoronzhuk@linaro.org>