ethernet: ti: am65-cpsw-qos: add EST fetch ram configuration
The iproute2 interface allows to configure EST cycle setting while
another one is still operational, and next new cycle - admin one,
starts right after old one is completed. So, it's convenient way to
split fetch RAM on two and work in two buffer operation mode.
Each buffer has 64 locations that should be enough and configurations
are supposed to be toggled w/o harm at next timer cycle. But it's not
only case, that can be implemented. If number of commands is not
enough to cover requested intervals for schedules, the one buffer mode
is supposed to be used, but toggle can happen only if EST is stopped
before. The schedules time intervals can require more commands for
10Mb and 100Mb connection speeds, so both, one and two buffer should
be supported allowing a user to decide later, by info messages.

This patch adds only fetch ram buffer toggle machinery and doesn't
fill the fetch buffer content, when the decision is made two or one
buffer is used.

Signed-off-by: Ivan Khoronzhuk <ivan.khoronzhuk@linaro.org>