ethernet: ti: am65-cpsw-qos: add mac hold/release feature routine
No interface to enable mac hold/reales per port as a whole. The taprio,
for instance, is capable to set this flag per each schedule. As the hw
is capable to hold/release preeption per port, lets add ethtool flag
for now, that can be done in-flight, at least for debugging reason.
Later, when interface for this is clean it can be changed to new
interface. There is hope that newer h/w is able to do it per schdule,
but that's related to EST more than to IET as standalone feature.

Signed-off-by: Ivan Khoronzhuk <ivan.khoronzhuk@linaro.org>